// sineedge_front/src/app/settings/model/product-master.ts
export class ProductMaster {
//      ID: number;
//      ProductName: string;
//      ProductAbbr: string;

        product_id: any;
        name: any;
        abbr: any;
        createdon: any;
        fk_createdby : any;
        updatedon : any;
        isactive : any;
        fk_updatedby :any;
      
}